lider_project/requirements.txt
Django==4.1.5
psycopg2-binary==2.9.10
phonenumbers==9.0.3
django-phonenumber-field==8.0.0
protobuf==3.20.3
deeppavlov==1.1.1
transformers==4.30.2
torch==2.0.1
matplotlib==3.5.0
seaborn==0.11.2
scikit-learn==1.0.2
tqdm>=4.42.0,<4.65.0